Add OpenAI passthrough backend.
This commit adds an OpenAI passthrough backend. In order to do this, a
few minor tweaks to the Backend interface were. More significantly, the
OpenAI API handling had to be tweaked to allow some additional methods.
The backend operates as a standard backend, but uses a placeholder model
name ("passthrough") to avoid allocating one runner per OpenAI model.

I've added a few more methods (most notably the rest of the chat
completions API and the responses API), but not all methods yet because
many of the multimodal APIs return responses that we can't record.

Signed-off-by: Jacob Howard <jacob.howard@docker.com>