WiP: Implement basic memory estimation in scheduler
First pass implementation of memory estimation logic in model scheduler.
This change heavily relies on gguf-parser-go to calculate estimated peak
memory requirement for running inference with a given model. It adds
GetRequiredMemoryForModel() to the Backend interface to allow each
backend to deal with its config and calculate required memory usage
based on it.

Before merging:
* Add support for windows (CUDA)
* Add support for Linux (CUDA)
* Estimate RAM usage, especially when GPU offload is not configured
* Extend loader to keep track of RAM and VRAM (for now one VRAM space)

Signed-off-by: Piotr Stankiewicz <piotr.stankiewicz@docker.com>