tools: add an opencode Docker image that uses local DMR models
This commit adds a new demo / tool that demonstrates using opencode with
local models via Docker Model Runner. It accomplishes this using a
custom entrypoint that interfaces with Docker Model Runner to provide an
optimal configuration based on the system size.

It also includes a demo that extends the opencode image with some custom
toolchains and shows how it might be run in a real-world project.

Signed-off-by: Jacob Howard <jacob.howard@docker.com>