Add modelrunner to render when executing DMR
The modelrunner user needs to be a member of the render group for
Vulkan to work. Unfortunately render does not have a fixed GID, so we
can't reliably handle this in the Dockerfile. So have the CLI query the
hosts render GID and add it when starting the DMR container.

Signed-off-by: Piotr Stankiewicz <piotr.stankiewicz@docker.com>