runner: Make runtime memory checking opt-in
The current model memory requirement checking routines are limited to
llama.cpp, and because a separate library is used to do the estimation
it is likely to lag behind llama.cpp in terms of supported models. This
significantly complicates delivering a good experience for users of the
feature. So make it opt-in, for now.

Signed-off-by: Piotr Stankiewicz <piotr.stankiewicz@docker.com>